backend/pyproject.toml
[tool.poetry]
name = "placy"
version = "0.1.0"
description = "Placement dashboard for Students"
authors = ["pspiagicw <pspiagicw@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.88.0"
httpx = "^0.23.1"
uvicorn = "^0.20.0"
pymongo = "^4.3.3"
python-dotenv = "^0.21.0"
pyjwt = "^2.6.0"
email-validator = "^1.3.0"
sendgrid = "^6.9.7"
mongoengine = "^0.24.2"
python-dateutil = "^2.8.2"
passlib = "^1.7.4"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
pytest = "^7.2.0"
pdoc = "^12.3.0"
pre-commit = "^2.20.0"
pydocstyle = "^6.1.1"
toml = "^0.10.2"
faker = "^15.3.4"
isort = "^5.11.3"

[tool.poe.tasks]
test = "pytest"
run = ["doc", { shell = "python -m placy"}]
format = [ { shell = "black placy tests"}]
init = [ { shell = "poetry install"} , { shell = "poetry run pre-commit install"} , { shell = "poetry run pre-commit "} ]
doc = "pdoc -o docs placy"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
